API, Core: Move SQLViewRepresentation to API
